Torpedo Scanner: Scan a data file for torpedos!
You are given two files, a noisy area  (TestData.blf) filled with
possible threats, and the image of a torpedo (SlimeTorpedo.blf).

The aim: Scan the area and identify if there are any torpedos.

However, the area is so noisy, a torpedo may not look the same as that
in the SlimeTorpedo file(pixels missing or added), so you need to
provide a rating as to how similar items are to a torpedo.